Initial support for a configuration file
Now supports an rcfile (defaults to ~/.galileorc) which can be used to
set all configuration options. Command-line arguments can override
these rcfile defaults, which is why both 'XXX' and 'no-XXX' versions
of flag arguments are now needed (eg both '--dump' and '--no-dump').

Configuration is built up in a dict, which is then used to pass all
configuration to the main sync method.